+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 - [#5728](https://github.com/influxdata/influxdb/issues/5728): Properly handle semi-colons as part of the main query loop.
 - [#6065](https://github.com/influxdata/influxdb/pull/6065):  Wait for a process termination on influxdb restart @simnv
 
-## v0.11.0 [unreleased]
+## v0.11.0 [2016-03-22]
 
 ### Release Notes